// client/cody/src/types.ts
export interface Message {
  speaker: 'you' | 'bot';
  text: string;
}
// ...
export interface WSChatMessage {
  kind: 'request' | 'response:change' | 'response:complete' | 'response:error';
  requestId: number;
}
export interface WSChatRequest extends WSChatMessage {
  kind: 'request';
  messages: Message[];
}
export type WSChatResponse =
  | WSChatResponseChange
  | WSChatResponseComplete
  | WSChatResponseError;
export interface WSChatResponseChange extends WSChatMessage {
  kind: 'response:change';
  message: string;
}
export interface WSChatResponseComplete extends WSChatMessage {
  kind: 'response:complete';
  message: string;
}
export interface WSChatResponseError extends WSChatMessage {
  kind: 'response:error';
  error: string;
}